# Lazy loaded shiki languages during syntax highlighting