--- a/.changeset/real-rabbits-bake.md
+++ b/.changeset/real-rabbits-bake.md
@@ -1,0 +1,7 @@
+---
+"@astrojs/markdown-remark": major
+---
+
+This changes the `markdown-remark` package to lazily load shiki languages by
+default (only preloading `plaintext`). Additionally, highlighting is now an
+async task due to this.