--- a/packages/markdown/remark/src/rehype-prism.ts
+++ b/packages/markdown/remark/src/rehype-prism.ts
@@ -3,12 +3,14 @@ import type { Root } from 'hast';
 import type { Plugin } from 'unified';
 import { highlightCodeBlocks } from './highlight.js';
 
-export const rehypePrism: Plugin<[], Root> = () => (tree) => {
-	highlightCodeBlocks(tree, (code, language) => {
-		let { html, classLanguage } = runHighlighterWithAstro(language, code);
+export const rehypePrism: Plugin<[], Root> = () => {
+	return async (tree) => {
+		await highlightCodeBlocks(tree, (code, language) => {
+			let { html, classLanguage } = runHighlighterWithAstro(language, code);
 
-		return Promise.resolve(
-			`<pre class="${classLanguage}"><code is:raw class="${classLanguage}">${html}</code></pre>`
-		);
-	});
+			return Promise.resolve(
+				`<pre class="${classLanguage}"><code is:raw class="${classLanguage}">${html}</code></pre>`
+			);
+		});
+	};
 };