… packages/create-astro/test/fixtures/not-empty/package.json