@43081j 2024-03-30
making this a draft for now as its possible we can't have these nice things unless rehype lets us do async 👀