@bluwy 2024-04-01
Thanks! All these changes make sense to me. About shiki lazy load, we actually used to do the same but I removed it because I thought shikiji/shiki 1.0 already did that by default, but in hindsight it probably didn't because its highlighter functions are all sync anyways (which you pointed out).

About your message on Discord on version bumps, this will require a major for `@astrojs/markdown-remark` as it's published. Adding a major changeset for it should be enough and changesets will auto bump the rest (I hope). There's already many things within `@astrojs/markdown-remark` that had been marked deprecated, and this seems like the right time to clean it all up.

Would you be able to add a major changeset for`@astrojs/markdown-remark`? I'll handle the other major changes in another PR.