@43081j 2024-04-01
have fixed the broken test

FYI it was working because of this it seems:
https://github.com/withastro/astro/blob/0be26d8b10fb19efaa9c1d2f88a4950d9550994a/packages/astro/src/runtime/server/escape.ts#L99-L103

i suppose `<Fragment set:html=${promise}>` results in it awaiting the promise internally via the above, so it was working fine